# feat: revamp feed history page with advanced filtering, pagination, statistics dashboard and comprehensive performance optimization